Despite terrible embarrassment at the state of this code, I'm going to put it up on github.
The idea was to build a demo app, but I'm a bit short on time, so the important parts've all been chucked into misc:

Misc dir:
---------------------------
file:   geochecker.lisp       (contains a claims-geochecker widget)
file:   maps.js
file:   older-geochecker.lisp (older-geochecker shows a previous, failed approach)

maps.js is loaded as a dependency of the geochecker widget which is called claims-geochecker and is defined in geochecker.lisp

Any field which has a rel="map" attribute will trigger the map when its text changes.

There was a version earlier that would show the map-dialog after clicking a button, but this way is better, imho.

There was also a version that built an accordion to categorize the various place-names that geonames returns (eg Place, Country etc)

I also investigated cl-geonames, but it was more user-friendly to build the UI in JS, and so ended up using the ajax option for the query.